requirements.txt
types-python-dateutil
python-dateutil
pytz
types-pytz
feedparser
genrss